refactor and add two stage eval scaffold
