+++ libeval/src/lib.rs
@@ -1,7 +1,11 @@
 pub mod actor;
 pub mod attribute;
+pub mod error;
 pub mod eval;
+pub mod eval_result;
 pub mod joinpolicy;
 pub mod logging;
 pub mod pio;
 pub mod policy;
+pub mod route;
+pub mod visa;
